Added support classes do make pattern matching
- Assignment
- QueryAnswer

Assignment holds a set of mappings label -> value determining a given
assignment of values to variables. In this initial version, only a simple
merge() operator were provided but the algebra of Assignments should be
extended as we port extend the pattern matcher to operates over logical
expressions (exporession wirth logical operators like AND, OR and NOT).

QueryAnwer is a dataclass to hold a given assignment, which is supposed
to be a valid answer to a query, and the respective grounded subgraph
that results from the applying the values given in the assignment to
the pattern given as query.